cambios en clase ResponseBean

+++ src/main/java/es/blaster/miniapicalc/ResponseBean.java
@@ -4,25 +4,25 @@ package es.blaster.miniapicalc;
 
 public class ResponseBean {
 
-    private int operando1;
-    private int operando2;
+    private double operando1;
+    private double operando2;
     private String operador;
     private double resultado;   
     private String ErrorDescription;
 
-    public int getOperando1() {
+    public double getOperando1() {
         return operando1;
     }
 
-    public void setOperando1(int operando1) {
+    public void setOperando1(double operando1) {
         this.operando1 = operando1;
     }
 
-    public int getOperando2() {
+    public double getOperando2() {
         return operando2;
     }
 
-    public void setOperando2(int operando2) {
+    public void setOperando2(double operando2) {
         this.operando2 = operando2;
     }
